refactor: extract src/indexer/infer/lambda.rs
Move pure lambda-type helper functions out of the indexer.rs monolith
into a new src/indexer/infer/lambda.rs submodule.

Functions moved (all pure: &str -> Option<String>, no Indexer dep):
  - SCOPE_FUNCTIONS constant
  - lambda_type_first_input
  - lambda_type_nth_input
  - lambda_type_receiver
  - strip_suspend (private helper, only used by the above)

New files:
  - src/indexer/infer/mod.rs      -- facade with pub(crate) re-exports
  - src/indexer/infer/lambda.rs   -- the moved pure functions
  - src/indexer/infer/lambda_tests.rs -- extracted unit tests (via #[path])

src/indexer.rs declares `mod infer;` and re-imports the symbols via
`use self::infer::{...};` so all existing call-sites remain unchanged.

Tests: 346 + 2 = 348 pass (unchanged count; the 2 moved tests now live
in indexer::infer::lambda::tests but count in the same binary).

Part of p1-lambda step in the refactor/split-modules series.
